+++ src/components/layout/QueryLayout.tsx
@@ -1,5 +1,10 @@
-import { useApiClient } from "@/core/hooks/api-client";
-import type { QueryLayoutProps } from "./types";
+import { useApiClient } from "@hooks/useApiClient";
+import type { QueryClientProps } from "@type/core/queryClient";
+
+type QueryLayoutProps<T> = {
+    queryClientOptions: QueryClientProps;
+    children: (data: T) => React.ReactNode;
+};
 
 export default function QueryLayout<T>(props: QueryLayoutProps<T>) {
     const { data, isLoading, isError, error } = useApiClient<T>(