+++ src/components/layout/QueryLayout/types.ts
@@ -1,6 +1,0 @@
-import type { QueryClientProps } from "@/core/types/api-client";
-
-export type QueryLayoutProps<T> = {
-  queryClientOptions: QueryClientProps;
-  children: (data: T) => React.ReactNode;
-};