+++ src/components/ui/InputText.tsx
@@ -1,4 +1,4 @@
-import type { InputTextProps } from "../types/input";
+import type { InputTextProps } from "@type/core/input";
 
 export default function InputText(props: InputTextProps) {
     const { value, placeholder, onChange, ariaLabel, type } = props;