+++ CHANGELOG.md
@@ -29,16 +29,17 @@
 ### ⚙️ Miscellaneous Tasks
 
 - Project scaffolding
-- *(vite-config)* Removed tanstack router plugin
+- _(vite-config)_ Removed tanstack router plugin
 - Removed unused package
 - Configured environment
 - Package.json syncronization on release
 - Added testing-library
 - Added jsom dependency
 - Configured alias for tests folder
+
 ## [0.1.0] - 2025-11-04
 
 ### ⚙️ Miscellaneous Tasks
 
 - Project scaffolding
-- *(vite-config)* Removed tanstack router plugin
+- _(vite-config)_ Removed tanstack router plugin